pyproject.toml
[project]
name = "bot2"
version = "0.1.0"
description = ""
authors = [
    {name = "AntonNovozhilov",email = "wowlass63@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram (>=3.21.0,<4.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
